Redesign `Desc` to deal with the fact that `JObject` and friends are now `!Copy`.
Note that it is now sometimes necessary to use turbofish syntax with `Desc::lookup`, since the return type of `Desc::<T>::lookup` is not necessarily `T` any more (it might be `&T` or something instead). For example, if you need a `JThrowable` and the compiler can't figure that out, write `Desc::<JThrowable>::lookup(input, env)` instead of `input.lookup(env)`. User code doesn't usually use `Desc` directly, so this shouldn't be too much of a problem.

This new design has the side benefit that looked-up values can be wrapped in `AutoLocal`, which fixes #109.